Create E-commerce Consumer Dataset
In this project, we work with a dataset available on Kaggle. The data includes information about customers of an e-commerce website, comprising the following:

Avg. Session Length: Average duration of in-store style advice sessions.
Time on App: Average time spent on the app in minutes.
Time on Website: Average time spent on the website in minutes.
Length of Membership: Number of years the customer has been a member.